Python: CG trace: Make code modular
