+++ python/tools/recorded-call-graph-metrics/.gitignore
@@ -1,0 +1,7 @@
+# Example DB
+cg-trace-example-db/
+
+# Artifact from building `pip install -e .`
+cg_trace.egg-info/
+
+venv/
