+++ python/tools/recorded-call-graph-metrics/README.md
@@ -14,10 +14,40 @@ The next hurdle is being able to handle multiple calls on the same line, such as
 
 ## How do I give it a spin?
 
-Run the `recreate-db.sh` script to create the database `cg-trace-example-db`. Then run the queries inside the `ql/` directory.
+After following setup instructions below, run the `recreate-db.sh` script to create the database `cg-trace-example-db`. Then run the queries inside the `ql/` directory.
 
 
-## Limitations
+## Setup
+
+1. Ensure you have at least Python 3.6
+
+2. Create virtual environment `python3 -m venv venv` and activate it
+
+3. Install dependencies `pip install -r --upgrade requirements.txt`
+
+4. Install this codebase as an editable package `pip install -e .`
+
+5. Setup your editor. If you're using VS Code, create a new project for this folder, and
+   use these settings for correct autoformatting of code on save:
+  ```
+  {
+      "python.pythonPath": "venv/bin/python",
+      "python.linting.enabled": true,
+      "python.linting.flake8Enabled": true,
+      "python.formatting.provider": "black",
+      "editor.formatOnSave": true,
+      "[python]": {
+          "editor.codeActionsOnSave": {
+              "source.organizeImports": true
+          }
+      }
+  }
+  ```
+
+6. Enjoy writing code, and being able to run `cg-trace` on your command line :tada:
+
+
+## Tracing Limitations
 
 ### Code that uses `sys.setprofile`
