+++ python/tools/recorded-call-graph-metrics/cg_trace/__init__.py
@@ -1,0 +1,15 @@
+import sys
+
+__version__ = "0.0.1"
+
+# Since the virtual machine opcodes changed in 3.6, not going to attempt to support
+# anything before that
+MIN_PYTHON_VERSION = (3, 6)
+MIN_PYTHON_VERSION_FORMATTED = ".".join(str(i) for i in MIN_PYTHON_VERSION)
+
+if not sys.version_info[:2] >= MIN_PYTHON_VERSION:
+    sys.exit(
+        "You need at least Python {} to use 'cg_trace'".format(
+            MIN_PYTHON_VERSION_FORMATTED
+        )
+    )
