+++ python/tools/recorded-call-graph-metrics/cg_trace/__main__.py
@@ -1,0 +1,5 @@
+import sys
+
+from cg_trace.main import main
+
+sys.exit(main())
